requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
python-dotenv>=0.21.0
diagrams>=0.23.0
graphviz>=0.20.0